requirements.txt
pandas==2.2.2
tqdm==4.66.4
geoip2==4.8.0
mysql-connector-python==8.4.0
